requirements.txt
### Package requirements for trading server ###`
alpha-vantage==2.3.1
finnhub-python==2.4.0
falcon==3.0.0
gunicorn==20.1.0
